Unlock timed out disabled jobs
Some apps depends on knowing when a job has finished running. It's possible for a job to be disabled, and then have the server crash and restart, and not be able to tell if a job is still running, or is timed out. This fix clears the lock and fixes that ambiquity.